[Android] Renamed IsEncryptEverythingAllowed and IsUsingSecondaryPassphrase
Chromium changes:

https://chromium.googlesource.com/chromium/src/+/930e2af1d16616915d2190503be1ccee143e1b53

[sync] Rename term 'secondary' with 'explicit' wrt passphrase

No behavioral changes, it's a fully automated replacement of the term.

The new name is consistent with other code like
syncer::IsExplicitPassphrase(). This logic returns true mainly for
custom passphrase users but also for the legacy
frozen-implicit-passphrase users (very similar to custom passphrase).

Change-Id: Id7ccbfaeef8fa995ca669e6040ef276bd18c0d3b
Bug: 1016715

https://chromium.googlesource.com/chromium/src/+/adfb75d061f224d12b009e31f4fc10b7da964868

[sync] Rename IsEncryptEverythingAllowed to IsCustomPassphraseAllowed

No behavior is changed. This better reflects the meaning of this
method.

Bug: None