speedreader: readability: Add a fast URL heuristic
Put a super lightweight URL heuristic on the speedreader readability
backend. This prevents us from running the expensive scorer (up to 30ms)
on a large amount of sites. The URL rule regexes are cached in a
singleton, so regex is only compiled once per browser run.

TODO: As a follow-up to this PR, we plan to include a large list of
"maybe-speedreadable" domains to work with popular websites that don't
meet these heuristics.

Resolves https://github.com/brave/brave-browser/issues/15809